[hip][cuda] Update event allocation and collection.
The existing system was not sufficient for graphs,
as they can be run out of order and have different
behavior for event recording.

This does not entirely solve the problem for re-use
if we ever want to simultaneously submit more than one graph
at a time. However this can be extended to do so in
the future.

Signed-off-by: Andrew Woloszyn <andrew.woloszyn@gmail.com>