Some optimizations for external events.
This allows us to return immediately from the any
async calls instead of having to wait for the dispatch
thread. The downside here is that it means the signal
goes across yet another stream. However this can
significantly reduce the amount of time we are on the
main thread. We also don't pay much of a cost unless a
semaphore actually exports an event.

Signed-off-by: Andrew Woloszyn <andrew.woloszyn@gmail.com>